[board-server] Remove logic from boards/index
Pushes all logic down into the handlers themselves.
