+++ packages/board-server/src/server/boards/invoke.ts
@@ -4,19 +4,27 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type { Response } from "express";
+import type { Request, Response } from "express";
+
+import { getBody } from "../common.js";
+import type { ServerConfig } from "../config.js";
+import { secretsKit } from "../proxy/secrets.js";
 
 import { loadFromStore } from "./utils/board-server-provider.js";
-import { verifyKey } from "./utils/verify-key.js";
-import { secretsKit } from "../proxy/secrets.js";
 import { invokeBoard } from "./utils/invoke-board.js";
+import { verifyKey } from "./utils/verify-key.js";
 
 async function invokeHandler(
-  url: URL,
-  res: Response,
-  body: unknown
+  config: ServerConfig,
+  req: Request,
+  res: Response
 ): Promise<void> {
   const { user, name, fullPath } = res.locals.boardId;
+  const url = new URL(req.url, config.hostname);
+  url.pathname = `boards/${fullPath}`;
+  url.search = "";
+
+  const body = await getBody(req);
   const inputs = body as Record<string, any>;
   const keyVerificationResult = await verifyKey(user, name, inputs);
   if (!keyVerificationResult.success) {
