+++ packages/board-server/src/server/boards/post.ts
@@ -7,11 +7,32 @@
 import type { Request, Response } from "express";
 
 import { ok, type GraphDescriptor } from "@google-labs/breadboard";
+
 import { authenticateAndGetUserStore } from "../auth.js";
 import { badRequest } from "../errors.js";
 import { getStore } from "../store.js";
+import { getBody } from "../common.js";
 
-async function post(req: Request, res: Response, body: unknown): Promise<void> {
+import del from "./delete.js";
+
+async function post(req: Request, res: Response): Promise<void> {
+  const body = await getBody(req);
+
+  // We handle deletion by accepting a POST request with { delete: true } in the body
+  // TODO Don't do this. Use HTTP DELETE instead.
+  const maybeDelete = body as { delete: boolean };
+  if (maybeDelete.delete === true) {
+    await del(req, res);
+  } else {
+    await update(req, res, body);
+  }
+}
+
+async function update(
+  req: Request,
+  res: Response,
+  body: unknown
+): Promise<void> {
   const boardPath = res.locals.boardId.fullPath;
   let store;
